# Conflicts:
#	src/main/java/com/example/projekt2_gruppe4_onskeskyen/repository/WishlistRepository.java
#	src/main/java/com/example/projekt2_gruppe4_onskeskyen/service/WishlistService.java
#	src/main/resources/templates/showWishlists.html

+++ src/main/java/com/example/projekt2_gruppe4_onskeskyen/controller/UserController.java
@@ -82,6 +82,14 @@ public class UserController {
     public String deleteUser(HttpSession session) {
         User user = (User) session.getAttribute("loggedInUser");
 
+        if (user == null) {
+            return "redirect:/login";
+        }
+
+        if (user.getId() <= 0) {
+            return "redirect:/";
+        }
+
         userService.deleteUser(user.getId());
 
         session.invalidate();
